+++ install.py
@@ -1,5 +1,6 @@
 #!/usr/bin/python3
 
+
 # Imports
 import subprocess
 import os